BLaZeeeeeeee UWU
mmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmm we suuureeee did build Blaze today 👼 isnt that sooooooo cool?
Man oh man I can't wait until Austin isn't looking to move every motor into the intake system and rename them all with different naming conventions, maybe for one motor ill call is INtaKeLeft and itll actually control the right ⏰.
Isn't that funny?
Anyways I added all the code for the competition Blaze (except for a bunch of things so i guess really this is just the bare minimum to make Blaze work and then Austin (Digi) can deal with the rest later 💯

+++ include/devils/geometry/vector2.hpp
@@ -110,7 +110,7 @@ namespace devils
             const float deltaX = x - other.x;
             const float deltaY = y - other.y;
 
-            return std::sqrtf(
+            return std::sqrt(
                 deltaX * deltaX +
                 deltaY * deltaY);
         }
@@ -121,7 +121,7 @@ namespace devils
          */
         float magnitude() const
         {
-            return std::sqrtf(
+            return std::sqrt(
                 x * x +
                 y * y);
         }
